Add request handler for /share
/share end-point is used as a web share target, and files can be shared with
the app from Android phones. This view uploads all the files to a temporary
directory on Google Drive, and then prompts the user to rename the
directory, or move the files to another directory on the drive.

Specifying the GDrive directory before the upload using service workers may
be a better approach, but is quite tricky to develop and debug. The approach
used in this commit is not the most user friendly, but is good enough for a
quick prototype.