Add information about uploader to the description of photos
Since all the photos are uploaded as the same user, it would be useful to
have information about who uploaded a picture, somewhere.

It would probably be nice to have this information in properties, in case
someone wants to do something with this information in the future, but this
information is not visible in the Google Drive UI. So, we just add the
information to the image description, currently.

Closes #37