splitwise: Improve syncing of RSVP event with Splitwise group
The API call now uses an API token instead of trying to do the sync as the
currently logged in user.  This makes the sync work more reliably and makes the
code simpler, because we don't need to worry if the current user is authorized
with Splitwise or has permissions to tweak the group, linked to the event.