models: Remove unique user restriction to fix broken indexes
There's a bunch of events which have duplicate user RSVPs because of the
anonymous user and one event with accidentally multiple RSVPs for a same user.

For now, we remove the unique index, but we could potentially create a bunch of
anonymous users and use them in the RSVPs.  This could also help make it easier
to invite guests to an event.

```python
from collections import Counter

for event in Event.objects.filter():
    emails = [rsvp.user.id for rsvp in event.rsvps]
    if len(set(emails)) != event.rsvps.count():
        print(event.id, Counter(emails), event.date)
        print("#" * 40)
```