splitwise: Fix incorrect grant_type usage of client_credentials
The client_credentials grant_type returns credentails for the owner of the app,
not the currently logged in user. Also, the session authorization doesn't seem
to work correctly. So, this commit simplifies the `/splitwise/allow` method to
simply ask for authorization and the post authorization hook handles the saving
of splitwise_id and adding the token admin as a friend.